Prevent duplicate links in REST post responses
